# packages/notifycore/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NotifyCore",
    "version": "0.1.0",
    "description": "Reliable notification delivery: templates, preferences, quiet hours, retries, DLQ, idempotency."
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header"
      }
    },
...
  },
  "paths": {
    "/notifications": {
      "post": {
        "summary": "Enqueue a notification (idempotent)",
        "tags": [
          "notifications"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idempotencyKey": {
                    "type": "string",
                    "minLength": 1
                  },
                  "recipient": {
                    "type": "string",
                    "minLength": 1
                  },
                  "channel": {
                    "type": "string",
                    "enum": [
                      "email",
                      "sms",
                      "push",
                      "webhook"
                    ]
                  },
                  "templateKey": {
                    "type": "string",
                    "minLength": 1
                  },
                  "data": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "maxAttempts": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0
                  }
                },
                "required": [
                  "idempotencyKey",
                  "recipient",
                  "channel",
                  "templateKey"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "idempotencyKey": {
                      "type": "string"
                    },
                    "recipient": {
                      "type": "string"
                    },
                    "channel": {
                      "type": "string"
                    },
                    "templateKey": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "status": {
                      "type": "string"
                    },
                    "attempts": {
                      "type": "number"
                    },
                    "maxAttempts": {
                      "type": "number"
                    },
                    "lastError": {
                      "type": "string",
                      "nullable": true
                    },
                    "scheduledFor": {
                      "type": "string",
                      "nullable": true
                    },
                    "sentAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "created": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "idempotencyKey",
                    "recipient",
                    "channel",
                    "templateKey",
                    "data",
                    "status",
                    "attempts",
                    "maxAttempts",
                    "lastError",
                    "scheduledFor",
                    "sentAt",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "idempotencyKey": {
                      "type": "string"
                    },
                    "recipient": {
                      "type": "string"
                    },
                    "channel": {
                      "type": "string"
                    },
                    "templateKey": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "status": {
                      "type": "string"
                    },
                    "attempts": {
                      "type": "number"
                    },
                    "maxAttempts": {
                      "type": "number"
                    },
                    "lastError": {
                      "type": "string",
                      "nullable": true
                    },
                    "scheduledFor": {
                      "type": "string",
                      "nullable": true
                    },
                    "sentAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "created": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "idempotencyKey",
                    "recipient",
                    "channel",
                    "templateKey",
                    "data",
                    "status",
                    "attempts",
                    "maxAttempts",
                    "lastError",
                    "scheduledFor",
                    "sentAt",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List notifications",
        "tags": [
          "notifications"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "queued",
                "sending",
                "sent",
                "deferred",
                "skipped",
                "dead"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "recipient",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0,
              "maximum": 500,
              "default": 50
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "idempotencyKey": {
                        "type": "string"
                      },
                      "recipient": {
                        "type": "string"
                      },
                      "channel": {
                        "type": "string"
                      },
                      "templateKey": {
                        "type": "string"
                      },
                      "data": {
                        "type": "object",
                        "additionalProperties": {}
                      },
                      "status": {
                        "type": "string"
                      },
                      "attempts": {
                        "type": "number"
                      },
                      "maxAttempts": {
                        "type": "number"
                      },
                      "lastError": {
                        "type": "string",
                        "nullable": true
                      },
                      "scheduledFor": {
                        "type": "string",
                        "nullable": true
                      },
                      "sentAt": {
                        "type": "string",
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      },
                      "created": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "id",
                      "idempotencyKey",
                      "recipient",
                      "channel",
                      "templateKey",
                      "data",
                      "status",
                      "attempts",
                      "maxAttempts",
                      "lastError",
                      "scheduledFor",
                      "sentAt",
                      "createdAt",
                      "updatedAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/notifications/{id}": {
      "get": {
        "summary": "Get a notification by id",
        "tags": [
          "notifications"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "idempotencyKey": {
                      "type": "string"
                    },
                    "recipient": {
                      "type": "string"
                    },
                    "channel": {
                      "type": "string"
                    },
                    "templateKey": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "status": {
                      "type": "string"
                    },
                    "attempts": {
                      "type": "number"
                    },
                    "maxAttempts": {
                      "type": "number"
                    },
                    "lastError": {
                      "type": "string",
                      "nullable": true
                    },
                    "scheduledFor": {
                      "type": "string",
                      "nullable": true
                    },
                    "sentAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "created": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "idempotencyKey",
                    "recipient",
                    "channel",
                    "templateKey",
                    "data",
                    "status",
                    "attempts",
                    "maxAttempts",
                    "lastError",
                    "scheduledFor",
                    "sentAt",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/templates": {
      "post": {
        "summary": "Create a template",
        "tags": [
          "templates"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string",
                    "minLength": 1
                  },
                  "channel": {
                    "type": "string",
                    "enum": [
                      "email",
                      "sms",
                      "push",
                      "webhook"
                    ]
                  },
                  "subject": {
                    "type": "string"
                  },
                  "body": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "key",
                  "channel",
                  "body"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "key": {
                      "type": "string"
                    },
                    "channel": {
                      "type": "string"
                    },
                    "subject": {
                      "type": "string",
                      "nullable": true
                    },
                    "body": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "key",
                    "channel",
                    "subject",
                    "body",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List templates",
        "tags": [
          "templates"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "key": {
                        "type": "string"
                      },
                      "channel": {
                        "type": "string"
                      },
                      "subject": {
                        "type": "string",
                        "nullable": true
                      },
                      "body": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "key",
                      "channel",
                      "subject",
                      "body",
                      "createdAt",
                      "updatedAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/templates/{key}": {
      "get": {
        "summary": "Get a template by key",
        "tags": [
          "templates"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "key",
            "required": true
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "key": {
                      "type": "string"
                    },
                    "channel": {
                      "type": "string"
                    },
                    "subject": {
                      "type": "string",
                      "nullable": true
                    },
                    "body": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "key",
                    "channel",
                    "subject",
                    "body",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Upsert a template",
        "tags": [
          "templates"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "channel": {
                    "type": "string",
                    "enum": [
                      "email",
                      "sms",
                      "push",
                      "webhook"
                    ]
                  },
                  "subject": {
                    "type": "string"
                  },
                  "body": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "channel",
                  "body"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "key",
            "required": true
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "key": {
                      "type": "string"
                    },
                    "channel": {
                      "type": "string"
                    },
                    "subject": {
                      "type": "string",
                      "nullable": true
                    },
                    "body": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "key",
                    "channel",
                    "subject",
                    "body",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}